Refactor  : 계산기 인터페이스 추가 및 상속
